Separate kintsugi migration from the roots migration
